optimistically load the editor while installing deps in background
