+++ content/package.json
@@ -14,28 +14,28 @@
     "@astrojs/check": "^0.9.4",
     "@astrojs/markdown-remark": "^5.3.0",
     "@astrojs/react": "^3.6.2",
-    "@astrojs/starlight": "^0.28.3",
+    "@astrojs/starlight": "^0.28.5",
     "@astrojs/starlight-tailwind": "^2.0.3",
     "@astrojs/tailwind": "^5.1.2",
     "@astrojs/vercel": "^7.8.2",
     "@dprint/formatter": "^0.4.1",
     "@effect-rx/rx-react": "^0.31.2",
-    "@effect/experimental": "^0.30.10",
+    "@effect/experimental": "^0.30.14",
     "@effect/monaco-editor": "^0.0.2",
-    "@effect/opentelemetry": "^0.39.4",
-    "@effect/platform": "^0.69.9",
-    "@effect/platform-browser": "^0.48.9",
-    "@effect/platform-bun": "^0.49.10",
-    "@effect/platform-node": "^0.64.10",
-    "@effect/rpc": "^0.44.9",
-    "@effect/rpc-http": "^0.42.9",
+    "@effect/opentelemetry": "^0.39.8",
+    "@effect/platform": "^0.69.13",
+    "@effect/platform-browser": "^0.48.13",
+    "@effect/platform-bun": "^0.49.14",
+    "@effect/platform-node": "^0.64.14",
+    "@effect/rpc": "^0.44.13",
+    "@effect/rpc-http": "^0.42.13",
     "@expressive-code/core": "0.35.6",
     "@expressive-code/plugin-collapsible-sections": "^0.35.6",
     "@expressive-code/plugin-line-numbers": "^0.35.6",
     "@opentelemetry/exporter-trace-otlp-http": "^0.54.0",
     "@opentelemetry/sdk-trace-base": "^1.27.0",
     "@radix-ui/react-alert-dialog": "^1.1.2",
-    "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/react-icons": "^1.3.1",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-toast": "^1.2.2",
     "@radix-ui/react-tooltip": "^1.1.3",
@@ -49,25 +49,25 @@
     "@webcontainer/snapshot": "^0.1.0",
     "@xterm/addon-fit": "^0.10.0",
     "@xterm/xterm": "^5.5.0",
-    "astro": "^4.16.7",
+    "astro": "^4.16.8",
     "astro-og-canvas": "^0.5.4",
     "astro-tweet": "^0.0.4",
     "canvaskit-wasm": "^0.39.1",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
-    "effect": "^3.10.4",
-    "framer-motion": "^11.11.10",
+    "effect": "^3.10.8",
+    "framer-motion": "^11.11.11",
     "hastscript": "^9.0.0",
-    "lucide-react": "^0.453.0",
-    "playwright": "^1.48.1",
+    "lucide-react": "^0.454.0",
+    "playwright": "^1.48.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-resizable-panels": "^2.1.5",
+    "react-resizable-panels": "^2.1.6",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-mermaid": "^3.0.0",
     "sharp": "^0.33.5",
     "starlight-blog": "^0.14.1",
-    "starlight-links-validator": "^0.12.3",
+    "starlight-links-validator": "^0.12.4",
     "tailwind-merge": "^2.5.4",
     "tailwindcss": "^3.4.14",
     "tailwindcss-animate": "^1.0.7",
@@ -86,6 +86,6 @@
     "@radix-ui/react-label": "^2.1.0",
     "@radix-ui/react-popover": "^1.1.2",
     "@radix-ui/react-tabs": "^1.1.1",
-    "cmdk": "1.0.0"
+    "cmdk": "1.0.3"
   }
 }
